rag/requirements.txt
# Core Framework
streamlit>=1.32.0
python-dotenv>=1.0.0

# PDF Processing
pypdf>=4.0.0
pdfplumber>=0.11.0

# Embeddings & Vector Store
sentence-transformers>=2.5.0
chromadb>=0.4.22
openai>=1.12.0

# NLP & Text Processing
tiktoken>=0.6.0
nltk>=3.8.1
spacy>=3.7.0

# Visualization
plotly>=5.19.0
matplotlib>=3.8.0
seaborn>=0.13.0
umap-learn>=0.5.5
scikit-learn>=1.4.0

# Utilities
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
pydantic>=2.6.0

# Optional: Local LLM support
# ollama>=0.1.0

# Optional: Advanced Vector DBs
# qdrant-client>=1.7.0
# pinecone-client>=3.0.0
